Обновление Add Arculator
Изменения:
- Добавлена игра arculator
- Переделана система файлов
