+++ Games/Arculator/saves
@@ -1,0 +1,2 @@
+SIZE: 0.7
+DIFFICULTY: 2
